Added ferret7 and ferret.py scripts, which enable traditional ferret command-line options; make fixes to ferret memory allocation similar to what was done in the trunk
git-svn-id: file:///home/users/tmap/svn/repos/ferret/branches/pyferret@11779 fdbf22ae-c210-0410-be80-ca943da6b8f8

+++ ferlib/ferret_lib.h
@@ -38,7 +38,7 @@ typedef enum CALTYPE_ {
 } CALTYPE;
 
 /* Prototypes for library C functions */
-void set_fer_memory(float *mem, int mem_size);
+void set_fer_memory(float *mem, size_t mem_size);
 void set_ppl_memory(float *mem, int mem_size);
 void set_shared_buffer(void);
 void decref_pyobj_(void *pyobj_ptr_ptr);
